requirements.txt
# BAMF-Eco: Budget-Aware Multi-Fidelity Eco-Optimizer
# NeurIPS 2026


# Core Deep Learning
torch>=2.5.0
torchvision>=0.20.0

# Object Detection Models (YOLOv8/v9/v10/v11/v26 + RT-DETR)
ultralytics>=8.4.0

# Bayesian Optimization
botorch>=0.17.0
gpytorch>=1.15.0

# Hyperparameter Optimization Baselines
hpbandster>=0.7.4
ConfigSpace>=1.2.0

# Scientific Computing
numpy>=2.0.0
scipy>=1.17.0
scikit-learn>=1.8.0
pandas>=3.0.0

# Visualization
matplotlib>=3.10.0
seaborn>=0.13.0

# GPU Power Monitoring
pynvml>=13.0.0

# Data & Evaluation
pycocotools>=2.0.6

# Config & Logging
PyYAML>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=9.0.0

# Stats
statsmodels>=0.14.0
